pyproject.toml
[tool.black]
line-length = 80

[tool.mypy]
check_untyped_defs = true
disable_error_code = 'attr-defined'
no_implicit_optional = true
show_error_codes = true
warn_unused_configs = true

[tool.poetry]
name = "kaist_menu"
version = "0.1.4"
description = ""
authors = ["bitsandscraps <17763502+bitsandscraps@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.0"
tomlkit = "^0.12.5"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
mypy = "^1.8.0"
black = "^24.3.0"
types-requests = "^2.31.0.20240106"
types-python-dateutil = "^2.9.0.20240906"

[tool.poetry.scripts]
kaistmenu = "kaist_menu:main_cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
